+++ scst/src/scst_targ.c
@@ -2181,7 +2181,7 @@ static void scst_do_cmd_done(struct scst_cmd *cmd, int result,
 
 	scst_set_exec_time(cmd);
 
-	WARN_ON_ONCE(IS_ERR_VALUE(result));
+	WARN_ON_ONCE(IS_ERR_VALUE((long)result));
 
 	cmd->status = result & 0xff;
 	cmd->msg_status = msg_byte(result);
